// projects/openapi-e2e-test/src/generated/book-store/responseServerTypes.ts
/*
 * This file has been generated by Oats, please don't modify it by hand!
 *
 * Generated from https://raw.githubusercontent.com/oats-ts/oats-schemas/master/schemas/book-store.json
 */

import { GetBooks200ResponseHeaderParameters } from './responseHeaderTypes'
import { AppError, Book } from './types'

export type AddBookServerResponse =
  | {
      statusCode: 201
      mimeType: 'application/json'
      body: Book
    }
  | {
      statusCode: 400
      mimeType: 'application/json'
      body: AppError[]
    }
  | {
      statusCode: 500
      mimeType: 'application/json'
      body: AppError[]
    }

export type GetBookServerResponse =
  | {
      statusCode: 200
      mimeType: 'application/json'
      body: Book
    }
  | {
      statusCode: 400
      mimeType: 'application/json'
      body: AppError[]
    }
  | {
      statusCode: 500
      mimeType: 'application/json'
      body: AppError[]
    }

export type GetBooksServerResponse =
  | {
      statusCode: 200
      mimeType: 'application/json'
      body: Book[]
      headers: GetBooks200ResponseHeaderParameters
    }
  | {
      statusCode: 400
      mimeType: 'application/json'
      body: AppError[]
    }
  | {
      statusCode: 500
      mimeType: 'application/json'
      body: AppError[]
    }