// projects/openapi-e2e-test/src/generated/book-store/types.ts
/*
 * This file has been generated by Oats, please don't modify it by hand!
 *
 * Generated from https://raw.githubusercontent.com/oats-ts/oats-schemas/master/schemas/book-store.json
 */

export type AppError = {
  message: string
}

export type Book = {
  author: string
  description?: string
  id: number
  price: number
  title: string
}